requirements.txt
streamlit==1.32.0
pandas==2.2.1
google-cloud-bigquery==3.18.0
db-dtypes==1.2.0
plotly==5.20.0